--- src/components/common/sidebar/dialog/room-creation-dialog.jsx
@@ -1,9 +1,0 @@
-function RoomCreationDialog() {
-  return (
-    <div>
-      test
-    </div>
-  );
-}
-
-export default RoomCreationDialog;
