--- src/components/common/sidebar/dialog/room-join-dialog.jsx
@@ -1,9 +1,0 @@
-function RoomJoinDialog() {
-  return (
-    <div>
-      test
-    </div>
-  );
-}
-
-export default RoomJoinDialog;
